refactoring taxonomy-search
