+++ inc/filter-search/package/src/taxonomy-search/index.js
@@ -1,0 +1,15 @@
+import { __ } from '@wordpress/i18n';
+import { ReactComponent as Icon } from './icon.svg';
+
+import metadata from './block.json';
+import edit from './edit';
+
+const { name } = metadata;
+
+export { metadata, name };
+
+export const settings = {
+	title: __( 'VK Taxonomy Search', 'vk-filter-search' ),
+	icon: <Icon />,
+	edit,
+};
